LayoutTests/platform/mac-bigsur-wk2/http/tests/websocket/tests/hybi/send-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/send' failed: The operation couldn’t be completed. (kNWErrorDomainPOSIX error 57 - Socket is not connected)
Web Socket send test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
PASS data is FIRST_MESSAGE_TO_SEND
PASS data is SECOND_MESSAGE_TO_SEND
Closed.
PASS successfullyParsed is true

TEST COMPLETE
